Add files via upload
This bar plot shows the effects that each feature has on the output of the model. We see that three of the features have large effects while the other two are much smaller. This implies that the model only considers three of the five features in its output and indeed this is how we designed the initial dataset. This algorithm can be applied to various machine learning models such as neural networks, random forest, logistic regression, k-nearest neighbors, etc.